requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
Deprecated==1.2.13
docker-pycreds==0.4.0
docstring-parser==0.15
filelock==3.8.0
frozenlist==1.3.1
fsspec==2022.10.0
gitdb==4.0.9
GitPython==3.1.29
google-auth==2.13.0
google-auth-oauthlib==0.4.6
grpcio==1.50.0
huggingface-hub==0.10.1
igraph==0.10.2
ijson==3.1.4
importlib-metadata==5.0.0
joblib==1.2.0
jsonargparse==4.15.2
lazy-import==0.2.2
Markdown==3.4.1
MarkupSafe==2.1.1
multidict==6.0.2
mypy-extensions==0.4.3
nltk==3.7
numpy==1.23.4
oauthlib==3.2.2
packaging==21.3
pandas==1.5.1
pathtools==0.1.2
Pillow==9.2.0
promise==2.3
protobuf==3.19.6
psutil==5.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyDeprecate==0.3.2
PyGithub==1.56
PyJWT==2.6.0
PyNaCl==1.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==1.7.7
pytz==2022.6
PyYAML==6.0
recordclass==0.17.5
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
sentry-sdk==1.10.1
setproctitle==1.3.2
seutil==0.8.6
seutils==1.16
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
texttable==1.6.4
tokenizers==0.13.1
torch==1.12.1
torchmetrics==0.10.0
torchvision==0.13.1
tqdm==4.64.1
transformers==4.23.1
typing-inspect==0.8.0
typing_extensions==4.4.0
unidiff==0.7.4
urllib3==1.26.12
wandb==0.13.4
Werkzeug==2.2.2
wrapt==1.14.1
xmltodict==0.13.0
yarl==1.8.1
zipp==3.9.0
